V-06 + F-04 1. Added reset of the search state when returning from the search results view page 2. Refactoring: search functions are moved to a separate file (searchManager.js)
